Fixed bug with requireAuth not working on routes. Issue was poor mounting implementation in the app.js
